requirements.txt
# ResNet-BK Requirements (Pinned)
# Python 3.10+

--extra-index-url https://download.pytorch.org/whl/cu118

# Core dependencies (GPU build, pinned for Triton compatibility)
torch==2.2.0+cu118
torchvision==0.17.0+cu118
torchaudio==2.2.0+cu118
triton==2.2.0; platform_system != "Windows"
ninja==1.11.1.1

# Hugging Face ecosystem
transformers==4.36.0
datasets>=2.16.1
tokenizers==0.15.0
accelerate==0.24.1
safetensors==0.4.0

# Scientific computing
numpy>=1.26.0  # Numerical stability (Py3.12 support)
scipy>=1.11.4  # Scientific computing
einops==0.7.0  # Tensor operations

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.1

# Utilities
tqdm>=4.66.0  # Progress bars
tqdm==4.66.1
pyyaml==6.0.1
rich==13.7.0
psutil==5.9.6

# Logging (optional)
tensorboard==2.15.1
wandb==0.16.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==6.1.0
mypy==1.6.1

# Jupyter (optional)
jupyter==1.0.0
jupyterlab==4.0.7
ipywidgets==8.1.1

# Phase 4 Dependencies
pyknotid>=0.5.0
tensornetwork==0.4.6
zarr==2.16.0
aiofiles==23.2.1
networkx==3.1
sympy==1.12

# UI / Dashboard
streamlit>=1.24.0
plotly>=5.18.0

# Rust Extension Build
maturin>=1.0.0
